Fixed header resizing issues / added functionality
