+++ Front end/src/components/update/Modal/header.scss
@@ -1,64 +1,90 @@
+/* Header.scss */
 .header {
-  background-color: #fff;
+  background-color: white;
   box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.2);
-  transition: transform 0.3s ease-in-out;
   position: fixed;
   top: 0;
   left: 0;
   right: 0;
+}
 
-  // Going to be used for whenever I implement the additional header features (This translates the header upward out of view
+.hamburger-menu {
+  display: flex;
+  justify-content: space-between;
+  flex-direction: column;
+  cursor: pointer;
+  margin-top: 40px;
+  margin-left: 20px;
+  margin-right: 80px;
+}
 
-  // &.menu-open {
-  //   transform: translateY(-100%);
-  // }
+.import-button-container {
+  position: relative;
+  width: 1px;
+  top: -39px;
+  left: 30px;
+  margin-left: 64px;
+  cursor: pointer;
 
-  .container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    max-width: 100%;
-    margin: 0 auto;
+  .import-button {
+    background-color: grey;
+    color: white;
+    border: none;
     padding: 10px 20px;
-  }
-
-  // Going to be used for whenever I implement the additional header features
-  .hamburger-menu {
-    display: flex;
-    flex-direction: column;
+    font-size: 16px;
     cursor: pointer;
-    padding: 5px;
+    border-radius: 4px;
 
-    .bar {
-      width: 30px;
-      height: 3px;
-      background-color: #000;
-      margin: 5px 0;
+    &:hover {
+        background-color: darkgray;
     }
-  }
-
-  .logo {
-    img {
-      max-width: 100%;
-      max-height: 100px;
-      display: block;
-      margin: 0 auto;
-    }
-  }
-
-  // Sidebar 
-.side-menu {
-  position: fixed;
-  top: 0;
-  left: -280px;
-  width: 280px;
-  height: 100%;
-  background-color: #333;
-  color: #fff; 
-  transition: left 0.3s ease-in-out; 
-
-  &.menu-open {
-    left: 0;
   }
 }
+
+.bar {
+  width: 30px;
+  height: 3px;
+  background-color: black;
+  margin: 5px 0;
+  cursor: pointer;
+}
+
+img {
+  max-width: 100%;
+  max-height: 130px;
+}
+
+.logo{
+  margin-top: -139px; 
+}
+
+.menu-open {
+  background-color: white;
+  box-shadow: 0px 10px 10px rgba(0, 0, 0, 0.2);
+  position: fixed;
+  width: 13%;
+  height: 100%;
+}
+
+.close-button {
+  z-index: 100000;
+  position: absolute;
+  top: 0;
+  right: 0;
+  margin-top: -5%;
+  font-size: 16pt;
+  background-color: red;
+  border-radius: 0px;
+  color: black;
+  padding: 4.5px;
+  cursor: pointer;
+}
+
+
+
+//This makes the DaiqUIris logo dissaper whenever the window width reaches 550px
+@media (max-width: 550px) {
+  .logo {
+    display: none;
+  }
 }
