--- Front end/src/components/update/Modal/importButton.tsx
@@ -1,25 +1,0 @@
-import React, { useState } from 'react';
-import './ImportButton.scss';
-
-const ImportButton = () => {
-    const [fileFinderVisible, setFileFinderVisible] = useState(false);
-
-    const toggleFileFinder = () => {
-        setFileFinderVisible(!fileFinderVisible);
-    };
-
-    return (
-        <div className="import-button-container">
-            <button className="import-button" onClick={toggleFileFinder}>
-                Import
-            </button>
-            {fileFinderVisible && (
-                <div className="fileFinder">
-                    <td>whatever file finder should open up</td>
-                </div>
-            )}
-        </div>
-    );
-};
-
-export default ImportButton;
